@@ -19,79 +19,99 @@
   ],
   "license": "MIT",
   "dependencies": {
-    "@types/bson": "1.x || 4.0.x",
-    "@types/mongodb": "^3.5.27",
-    "bson": "^1.1.4",
-    "kareem": "2.3.2",
-    "mongodb": "3.7.4",
-    "mongoose-legacy-pluralize": "1.0.2",
-    "mpath": "0.8.4",
-    "mquery": "3.2.5",
-    "ms": "2.1.2",
-    "optional-require": "1.0.x",
-    "regexp-clone": "1.0.0",
-    "safe-buffer": "5.2.1",
-    "sift": "13.5.2",
-    "sliced": "1.0.1"
+    "bson": "^6.7.0",
+    "kareem": "2.6.3",
+    "mongodb": "6.6.2",
+    "mpath": "0.9.0",
+    "mquery": "5.0.0",
+    "ms": "2.1.3",
+    "sift": "17.1.3"
   },
   "devDependencies": {
-    "@babel/core": "7.10.5",
-    "@babel/preset-env": "7.10.4",
-    "@typescript-eslint/eslint-plugin": "4.10.0",
-    "@typescript-eslint/parser": "4.10.0",
-    "@types/babel-types": "7.0.11",
-    "@types/babylon": "6.16.6",
-    "@types/json-schema": "7.0.3",
-    "@types/node": "16.11.23",
-    "acquit": "1.x",
-    "acquit-ignore": "0.1.x",
-    "acquit-require": "0.1.x",
-    "babel-loader": "8.1.0",
-    "benchmark": "2.1.4",
-    "bluebird": "3.5.5",
-    "chalk": "2.4.2",
-    "cheerio": "1.0.0-rc.2",
-    "co": "4.6.0",
-    "dox": "0.3.1",
-    "eslint": "7.1.0",
-    "eslint-plugin-mocha-no-only": "1.1.0",
-    "highlight.js": "9.18.2",
+    "@babel/core": "7.24.6",
+    "@babel/preset-env": "7.24.6",
+    "@typescript-eslint/eslint-plugin": "^6.2.1",
+    "@typescript-eslint/parser": "^6.2.1",
+    "acquit": "1.3.0",
+    "acquit-ignore": "0.2.1",
+    "acquit-require": "0.1.1",
+    "assert-browserify": "2.0.0",
+    "axios": "1.1.3",
+    "babel-loader": "8.2.5",
+    "broken-link-checker": "^0.7.8",
+    "buffer": "^5.6.0",
+    "cheerio": "1.0.0-rc.12",
+    "crypto-browserify": "3.12.0",
+    "dotenv": "16.4.5",
+    "dox": "1.0.0",
+    "eslint": "8.57.0",
+    "eslint-plugin-markdown": "^5.0.0",
+    "eslint-plugin-mocha-no-only": "1.2.0",
+    "express": "^4.18.1",
+    "fs-extra": "~11.2.0",
+    "highlight.js": "11.8.0",
     "lodash.isequal": "4.5.0",
     "lodash.isequalwith": "4.4.0",
-    "marked": "1.1.1",
-    "mkdirp": "0.5.5",
-    "mocha": "5.x",
+    "markdownlint-cli2": "^0.13.0",
+    "marked": "4.3.0",
+    "mkdirp": "^3.0.1",
+    "mocha": "10.4.0",
     "moment": "2.x",
-    "node-static": "0.7.11",
-    "object-sizeof": "1.3.0",
-    "pug": "2.0.3",
+    "mongodb-memory-server": "9.3.0",
+    "ncp": "^2.0.0",
+    "nyc": "15.1.0",
+    "pug": "3.0.3",
     "q": "1.5.1",
-    "rimraf": "2.6.3",
-    "schema-utils": "2.6.5",
-    "semver": "5.5.0",
-    "typescript": "4.1.x",
-    "uuid": "2.0.3",
-    "uuid-parse": "1.0.0",
-    "validator": "10.8.0",
-    "webpack": "4.44.0"
+    "sinon": "18.0.0",
+    "stream-browserify": "3.0.0",
+    "tsd": "0.31.0",
+    "typescript": "5.4.5",
+    "uuid": "9.0.1",
+    "webpack": "5.91.0"
   },
   "directories": {
     "lib": "./lib/mongoose"
   },
   "scripts": {
+    "docs:clean": "npm run docs:clean:stable",
+    "docs:clean:stable": "rimraf index.html && rimraf -rf ./docs/*.html  && rimraf -rf ./docs/api && rimraf -rf ./docs/tutorials/*.html && rimraf -rf ./docs/typescript/*.html && rimraf -rf ./docs/*.html && rimraf -rf ./docs/source/_docs && rimraf -rf ./tmp",
+    "docs:clean:5x": "rimraf index.html && rimraf -rf ./docs/5.x && rimraf -rf ./docs/source/_docs && rimraf -rf ./tmp",
+    "docs:clean:6x": "rimraf index.html && rimraf -rf ./docs/6.x && rimraf -rf ./docs/source/_docs && rimraf -rf ./tmp",
+    "docs:copy:tmp": "mkdirp ./tmp/docs/css && mkdirp ./tmp/docs/js && mkdirp ./tmp/docs/images && mkdirp ./tmp/docs/tutorials && mkdirp ./tmp/docs/typescript && mkdirp ./tmp/docs/api && ncp ./docs/css ./tmp/docs/css --filter=.css$ && ncp ./docs/js ./tmp/docs/js --filter=.js$ && ncp ./docs/images ./tmp/docs/images && ncp ./docs/tutorials ./tmp/docs/tutorials && ncp ./docs/typescript ./tmp/docs/typescript && ncp ./docs/api ./tmp/docs/api && cp index.html ./tmp && cp docs/*.html ./tmp/docs/",
+    "docs:copy:tmp:5x": "rimraf ./docs/5.x && ncp ./tmp ./docs/5.x",
+    "docs:copy:tmp:6x": "rimraf ./docs/6.x && ncp ./tmp ./docs/6.x",
+    "docs:generate": "node ./scripts/website.js",
+    "docs:generate:sponsorData": "node ./scripts/loadSponsorData.js",
+    "docs:test": "npm run docs:generate",
+    "docs:view": "node ./scripts/static.js",
+    "docs:prepare:publish:stable": "git checkout gh-pages && git merge master && npm run docs:generate",
+    "docs:prepare:publish:5x": "git checkout 5.x && git merge 5.x && npm run docs:clean:stable && npm run docs:generate && npm run docs:copy:tmp && git checkout gh-pages && npm run docs:copy:tmp:5x",
+    "docs:prepare:publish:6x": "git checkout 6.x && git merge 6.x && npm run docs:clean:stable && env DOCS_DEPLOY=true npm run docs:generate && mv ./docs/6.x ./tmp && git checkout gh-pages && npm run docs:copy:tmp:6x",
+    "docs:prepare:publish:7x": "env DOCS_DEPLOY=true npm run docs:generate && git checkout gh-pages && rimraf ./docs/7.x && mv ./tmp ./docs/7.x",
+    "docs:check-links": "blc http://127.0.0.1:8089 -ro",
     "lint": "eslint .",
-    "build-browser": "node build-browser.js",
+    "lint-js": "eslint . --ext .js --ext .cjs",
+    "lint-ts": "eslint . --ext .ts",
+    "lint-md": "markdownlint-cli2 \"**/*.md\"",
+    "build-browser": "(rm ./dist/* || true) && node ./scripts/build-browser.js",
     "prepublishOnly": "npm run build-browser",
     "release": "git pull && git push origin master --tags && npm publish",
-    "release-legacy": "git pull origin 5.x && git push origin 5.x --tags && npm publish --tag legacy",
-    "test": "mocha --exit ./test/*.test.js ./test/typescript/main.test.js",
-    "tdd": "mocha ./test/*.test.js ./test/typescript/main.test.js --watch --recursive --watch-files ./**/*.js",
-    "test-cov": "nyc --reporter=html --reporter=text npm test"
+    "release-5x": "git pull origin 5.x && git push origin 5.x && git push origin 5.x --tags && npm publish --tag 5x",
+    "release-6x": "git pull origin 6.x && git push origin 6.x && git push origin 6.x --tags && npm publish --tag 6x",
+    "mongo": "node ./tools/repl.js",
+    "publish-7x": "npm publish --tag 7x",
+    "test": "mocha --exit ./test/*.test.js",
+    "test-deno": "deno run --allow-env --allow-read --allow-net --allow-run --allow-sys --allow-write ./test/deno.js",
+    "test-rs": "START_REPLICA_SET=1 mocha --timeout 30000 --exit ./test/*.test.js",
+    "test-tsd": "node ./test/types/check-types-filename && tsd",
+    "tdd": "mocha ./test/*.test.js --inspect --watch --recursive --watch-files ./**/*.{js,ts}",
+    "test-coverage": "nyc --reporter=html --reporter=text npm test",
+    "ts-benchmark": "cd ./benchmarks/typescript/simple && npm install && npm run benchmark | node ../../../scripts/tsc-diagnostics-check"
   },
   "main": "./index.js",
-  "types": "./index.d.ts",
+  "types": "./types/index.d.ts",
   "engines": {
-    "node": ">=4.0.0"
+    "node": ">=16.20.1"
   },
   "bugs": {
     "url": "https://github.com/Automattic/mongoose/issues/new"
